Update Go Module Name
This patch updates the name of the Go module and the package paths to
"github.com/vmware-tanzu/service-apis". The CRD API group is retained,
but the module and package names must be updated so this repository can
continue to be used when the switch back to upstream occurs. Otherwise
it will not be possible to import the older versions in order to create
webhooks that convert to the newer versions.