Skip v1alpha0 TcpRoute/TrafficSplit
Do not generate CRDs for v1alpha0 TcpRoute and TrafficSplit. The former
is cased differently upstream, fixed in v1alpha1pre1, and the latter is
no longer part of the project.

If the old TcpRoute is generated, then the new version cannot be
generated because the new TCPRoute is cased different, and CRD kind
checks are case-insensitive.