Regen clientsets for CVE-2021-3121
This patch updates the tooling and Kubernetes dependencies to
a more recent version in order to account for the fix for
CVE-2021-3121.

This also means the clientsets have been regeenrated at
a newer version, and therefore existing integrations may break
as the generated clientsets may have function signatures that
have changed.

Note: boiler plate errors are expected, and haven't been updated.